+++ src/it/scala/com/ocadotechnology/pass4s/connectors/jms/JmsTests.scala
@@ -2,15 +2,14 @@ package com.ocadotechnology.pass4s.connectors.jms
 
 import cats.effect.IO
 import cats.effect.Resource
-import cats.implicits._
 import com.ocadotechnology.pass4s.connectors.activemq.Jms
 import com.ocadotechnology.pass4s.connectors.activemq.JmsDestination
 import com.ocadotechnology.pass4s.connectors.activemq.JmsSource
 import com.ocadotechnology.pass4s.connectors.activemq.JmsSource.JmsSourceSettings
-import com.ocadotechnology.pass4s.connectors.util.EmbeddedJmsBroker._
 import com.ocadotechnology.pass4s.core.Message
 import com.ocadotechnology.pass4s.high.Broker
 import com.ocadotechnology.pass4s.kernel.Consumer
+import com.ocadotechnology.pass4s.util.EmbeddedJmsBroker._
 import fs2.concurrent.SignallingRef
 import org.typelevel.log4cats.Logger
 import org.typelevel.log4cats.slf4j.Slf4jLogger
