+++ src/it/scala/com/ocadotechnology/pass4s/util/EmbeddedJmsBroker.scala
@@ -1,4 +1,4 @@
-package com.ocadotechnology.pass4s.connectors.util
+package com.ocadotechnology.pass4s.util
 
 import akka.actor.ActorSystem
